Add document filtering by standard type
The `documents` query returns all available documents at the
moment, but we need more control over it. One of the common
use case is to be able to filter document by type, and this
commit actually takes that into consideration and implements
it. Here is a sample query for that would only return one
type of documents.

```
{
  documents(type: "csd-standard") {
    id
    type
    bibdata {
      type
      status
    }
  }
}
```

Fixes: #9